… CACHING_ARCHITECTURE.md
… PERFORMANCE_IMPROVEMENTS.md
… PERFORMANCE_VALIDATION.md
… bin/main/config.yml
… src/main/java/net/lumalyte/game/GameCelebrationManager.java
… src/main/java/net/lumalyte/game/GameWorldManager.java
… src/main/java/net/lumalyte/gui/GameSetupMenu.java
… src/main/java/net/lumalyte/gui/TeamSelectionMenu.java
… src/main/java/net/lumalyte/hooks/HookManager.java
… src/main/java/net/lumalyte/listeners/AdminWandListener.java
… src/main/java/net/lumalyte/listeners/PlayerListener.java
… src/main/java/net/lumalyte/statistics/StatisticsDatabase.java
… src/main/java/net/lumalyte/util/ArenaWorldCache.java
… src/main/java/net/lumalyte/util/CacheManager.java
… src/main/java/net/lumalyte/util/ConcurrentChestFiller.java
… src/main/java/net/lumalyte/util/ConfigurationManager.java
… src/main/java/net/lumalyte/util/ErrorHandlingUtils.java
… src/main/java/net/lumalyte/util/GameInstancePool.java
… src/main/java/net/lumalyte/util/GuiComponentCache.java
… src/main/java/net/lumalyte/util/InvitationManager.java
… src/main/java/net/lumalyte/util/ItemUtils.java
… src/main/java/net/lumalyte/util/LootTableCache.java
… src/main/java/net/lumalyte/util/OptimizedConfigLoader.java
… src/main/java/net/lumalyte/util/PerformanceProfiler.java
… src/main/java/net/lumalyte/util/PlayerDataCache.java
… src/main/java/net/lumalyte/util/ScoreboardCache.java
… src/main/java/net/lumalyte/util/SkinCache.java
… src/main/java/net/lumalyte/util/StatisticsColumnMapper.java
… src/main/resources/config.yml
… src/test/java/net/lumalyte/performance/LumaSGPerformanceBenchmark.java